refactoring of storage decorators
